pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stellar-avoidance"
version = "1.0.0"
description = "SETI target avoidance framework for stellar catalogs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Şahin Torlakcık", email = "torlakciksahin@gmail.com"},
]
dependencies = [
    "numpy>=1.22",
    "pandas>=1.5",
    "astropy>=5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
stellar-avoidance = "stellar_avoidance.cli:main"

[tool.setuptools.packages.find]
include = ["stellar_avoidance*"]

[tool.setuptools.package-data]
stellar_avoidance = ["../configs/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
